fix(CustomSelect): Don't show clear button on option click with empty value
When CustomSelect or Select is controlled and value is '' with enabled
flag allowClearButton we show clear button after clicking on any option
without changing  props.value.
This is due to the fact that no option click we set nativeSelectValue
and clearButton we used to show by looking at nativeSelectValue,
but when controlled Select prop.value is not changed and still '' it is
wrong to show clearButton based on value of nativeSelectValue.

Steps to reproduce:
 - set value of controlled CustomSelect as ''
 - add allowClearButton flag to the component
 - don't change value passed to CustomSelect on change
 - click on any option from the Select dropdown.
 - you'll see clar button while value is still empty