Align CustomSelect behavior with NativeSelect
Add tests to cover expected behaviour
We explicitely fire onChange when
controlled component inner value is out of sync with
props.value.

It could happen when we don't change controlled value (props.value)
when some option is selected. In that case inner native select value
won't be the same as the props.value.
Select used to not fire change event when user clicks again on the same option
as native event value is already changed and stays the
same.
Now we help Select to trigger onChange in that situation.
We check that component is controlled, that we focused on the option
that is not selected yet and that props.value is out of sync with inner
native select value.