fix(PullToRefresh): Center spinner related to content
Spinner inside PullToRefresh is not centered if wrapper of PullToRefresh
has paddings, because Spinner is centered related
to FixedLayout component which is fixed, means window's width
is a primary width for the FixedLayout to calculate upon, not
a width of a wrapper or width of PullToRefresh.

To fix that we added `useParentWidth` prop to `FixedLayout` component.
When it's true we take the width of FixedLayout parent component,
ignoring css `width: 100%` which doesn't work for our case because
`FixedLayout` has `position: fixed` and `width` will be calculaed based on
window `width`.