Track usage of global tokens class using window
There are might be many independent apps those are using ConfigProvider
and those setup the tokens class on the body.
Then some of apps are unmounted ConfigProvider removes the class from
the body without checking whether there are any other apps those still
needs it.

Check that with VKUIGlobalTokensClassNameUsage class instance that
creates __VKUI__ variable in window and tracks the usage of className
on mount/unmount of ConfigProvider.

Now ConfigProvider removes the class only when the last app that is
using this class is unmounted.