Apply bottom-inset height in ModalPage transition calculation
Added new ref specially for bottom-inset to be able to get the
inset value by getting height of the element.

Use the bottom-inset value in content height during transition
calculation to make sure we move modal up not only by the height of the
content, but also by the height of the bottom-inset otherwise modal
content might be hidden by the bottom-inset.

Use this ref for new div element inside content-wrap.
Div element we need not only to use in calculation but also to limit
content scroll to make sure scroll visually ended before the
bottom-inset.