Add size:ci command to package.json of monorepo
To support migration to yarn v3 without fake yarn.lock file in
packages/vkui
We are going to drop `direction: ` option of `andresz1/size-limit-action`
because it's not going to allow us use this action without failure if we
remove fake yarn.lock file from packages/vkui

We have to remove fake yarn.lock file as yarn v3 doesn't install
dependencies properly with this fake file.

In the future PR we are going to use this new script in size-limit
action.