Remove extra padding of ActionSheet shadow on iOS
resolves: #3866

Shadow had extra padding because was applied to a warapper with padding
Unfortunately removing the padding is not enough because
there is also a space between content and cancel button in iOS,
and there we also see extra space of the shadow.

In order to avoid that we have to cover each part of ActionSheet
into wrapper and apply shadow on each part.