Don't set swipe back has tabbar to false
Firstly if we do it too early the overlay that
still has transition will cover the tabbar.
Secondly, yes, we can set state to false
when transition finished, but we don't really need it.
We use this state value only during the swiping back logic.
And every time user starts the transition we set it's value again.