Fix flacky expanding and scrolling
fixes: #5793
related: #4625, #4689, #5402

Sometimes it's possible to close the modal in a way
that doesn't turn on collapsed flag.
So, this flag is false and when it is off .ModalPage__content-in
has height: 100%.

We use .ModalPage__content-in offsetHeight value to compare
with it's parent's clientHeight to understand whether the modal content
is bigger than the modal height. If we have overflow - then we should
open the modal in expandable mode, which means to show only part
of the modal to allow user to expand it.
But when the .ModalPage__content-in height: 100% it means that it's
offsetHeight is pretty close to parent's clientHeight and when they are
equals we decided that this modal has no expandable mode and
we can show the full height modal.

Unfortunately, in that case we unexpectedly show the whole modal
which is not prepared for scrolling.

The solution here is to change the way we calculate the
height of the modal content.
We used to use clientHeight which depends on the css height,
so with height: 100% we wouldn't really get the real height of the
content with clientHeight, especially if the content is long.

Now we use scrollHeight which gives us real height of the content
so we can be accurate to decide whether the modal is expandable or not.