Fix scrolling behaviour in the collapsed state of modalPage
In the collapsed state modal shouldn't have an ability to scroll
But if we drag modal from expanded state to collapsed state from top to
bottom pretty fast, the we will be in the collapsed state visually, but
modalPage will treat the state as not collapsed and will allow scoroll
which is wrong.
Another case is when we are in the collapsed mode and drag top a little,
the scoll will be allowed too, while we still will be in collapsed mode
visually.

The reason behind this is the logic that make modalPage think that it is
in the collapsed mode.

```
modalState.collapsed = translateY > 0 && translateY < shiftYEndPercent;
```

And the problem mostly in shiftYEndPercent which can be less than
computed value of translateY in two casese above.

There are two problems with collapsed mode:
1. We use shiftYEndPercent to calculate collapsed flag
We shouldn't use it as it doesn't adjust it's value in case the app has hasCustomPanelHeaderAfter=true
or top inset. Ideally it should have values close to the translateY, but
with custom panel header values started to be different and logic is not
working as expected. The difference between translateY and
shiftYEndPerce is too big.

2. collapsedRange is defined even when the modal should not have it
For example when modal is not expanded. In that case modal is open
fully without having collapsed mode at all, but we define it's range
which is equals to expanded range and if we would use numberInRange(translateY, modalState.collapsedRange)
to calculate collapsed flag it would be true even when modal is expended
which will disable the scrolling behaviour.

So, the solution is to calculate collapsed flag using numberInRange(translateY, modalState.collapsedRange)
and disable collapsedRange when modal doesn't have collapsed state.

----

Remove extra css selector (defined in
851515bec376c44ed4e342229bb2b4ddb5874ef9) which makes 100% height of
.ModalPage__content-in when ModalPage state is not collapsed

We don't need anymore. Don't see any issues without it.