Detect pull to refresh gesture in the document's touchmove
Sometimes pullToRefresh triggers native page refresh insted of
component refresh.
The reason for it is that PullToRefresh set refreshing state too late
and native browser refresh might be triggered earlier.
In order to fix that we should cancel the native touchmove event once we
define that the touchmove is going to be a pull to refresh gesture
We do that by copying the logic we use in the Touch compoent onTouchMove
handler.
We check that page has no scroll offset and offsetY of touchmove event
is positive.