Split ScrollSaver to components
ScrollSaver is a component-wrapper which expects children
and search for the ref in the children props

ScrollSaverWithoutChildren is a component that receives ref from the
outside as elementRef prop to support dynamic scroll control.

useScrollSaver is a hook that allows to use scroll saver using hook.
Expects ref as a prop to control scroll position