Always show input if the Select is in searhable mode
Having it to jump from Input to SelectMimicry creates
issues with focus on select.
Focus is jumping out because on blur focused Input becomes SelectMimcry
and we lose focus.

Also, since now we put into the text input the selected option label.
And restore selected option label on blur or clear it if nothing is
selected.

We try to open the popup when we type in the input.
Also we trigger open state of popup when we type Backspace/Delete