Remove label wrapper from NativeSelect
Because it triggers screen reader to read the selected value as label
so it pronounce it twice in some versions of Chrome.
Anyway it will work properly even without label as native select
is streatched and above the visible content, so click on any area inside
NativeSelect is a click by native select.