Fix shouldDisablePortalLogic in AppRootPortal
After refactoring in c7e4c410eaf9877f236d5d7e1cfda49352197daa
We inversed the logic and forgot to inverse the mode condition too.
If we want to disablePortal in fallback we should disable the portal
when the mode='full'

That issue affected Snackbar position when it is used together with
Epic and tabbar.