Remove initialUpdate logic from floating Mutation observer callback
Remove that logic as it was added as optimization to avoid
re-rendering on intial load.
Turned out that MutationObserver callback is not called on initial load.

But this additional logic breaks positioning of filtered options
of ChipsSelect if it positioned on the bottom of the window.
In that case if user enter into the input text that filters all
options or that leaves one suggested option of the select then
the option will be rendered too hight compared to the select position.

This happens only when user enter the text that triggers rerender of
select options list for the first time.
And that is because of the initialUpdate logic.

Looks like unnecessary optimization.