Fix floating hook tests
Since FocusTrap is no longer depends on the keyboard interaction
the logic that prevents blur in specific cases is working
Like when we blur on the floating element: it treates like blur
due to click on the same element and we ignore it.

Before that change the test didn't event had the focus inside the
floating element, so, the blur worked in test.