Automatically apply tokens to html element
Plus automatically set .vkui on html and .vkui__root on app container

The reason for adding token class to the html element in
fact that tokens needs to be available on html element layer,
as we need them to set color-scheme property and html background.

Html background we need as in bounce effect when user overscrolls
we see animation and background of html element, which could be white
in dark mode otherwise, and even if we set color-scheme
html element will be dark, but no with out token color, but with default
browser agent style which is not appropriate.
Color-scheme needs to be set on html element as otherwise scroll bar
would still be white in dark mode.

We need this only in full mode as in embedded we didn't set this
previously either.

For propper SSR support user can set the token class on
server side, but should provide then proper appearance and platform
props, to avoid changes during hydration.