Remove ModalPopoutPortalRoot
It overcomplicates things, especially as
we still allow to pass modal/popout to SplitLayout
but it's no longer the requirenment.

---

During testing of prevoius implementation with rendering
directly to SlitLayout via portal turned out
that tests are not passing while nothing has really changed
for how we pass elements for portals.
Turned out tests relies on extra rerender of AppRoot,
so kind of on first automatic render we had a rerender
which in tests looks like portals are rendered directly to dom
on first glance.
It was not true, we needed extra render.
For modals it is via updating state to show modal.
Or via using state for refs, not React.useRef
because portals are not rendered on initial render as ref has not yet been set.