Get rid of options state
Use local options which we use as filtered data with useMemo

No more usage of options as state, since it takes at least one render
to update local option in useEffect and show options in the dropdown.
It's not working properly when we fetch options with opened dropdown.
Since it shows for a moment that there is no options, while they are
here, just local state is not updated.