Try approach with spinbutton role
This role is not recommended to use by MDN
Better to use native inputs, but it might be perfect fit
for our case when the majority of input functionality is
  done by InputLike component.

Fix focus with TAB/Shift + TAB
Open calendar by <Space> and always show calendar icon
Add FocusTrap to DateInput
Add role dialog to DateInput calendar
Add support for Home/End/PageUp/PageDown
  In Calendar days navigation
Add grid role to Calendar

Preserve focus state on focused day.
It means when user uses Tab to move away from
focused day on calendar we keep it's value,
so, when user goes back to it we focus exactly that
day.

Fix navigation with keyaboard through days

Allow to navigate to days using Tab
and focusable day is current value in the input,
or first day of the Month if month is changed,
but input value is not.

Allow NVDA to navigate via ctrl+alt+arrow

Restore focus on active day

Select date on Enter/Space key

No label with date if there is no value

Add day label without specific year, as it is too match info on calendar
to listen

Fix click outside logic

When the calendar is opened we don't have
selected value in the input
Should we have one?

Visually focus days when user clicks on label

Delay range selection to fix Firefox behavior

For some reason when we focus first or last
element of DateInput from the outside using Tab,
then we focus the first InputLike, but the content of
it is not selected for some reason.
The delay allows us to visually show the selection

Fix position of Calendar in doc Readme

It falls outside of iframe if we
disable iframe.
Add Panel and View and use iframe to
keep render it properly inside the doc

Add current month label to the table grid

Remove aria-modal="true"

Dialog is enough here.
We allow to interact with other element
before closing the Calendar

Deprecate changeDayLabel in Calendar

Since if follow MUI calendar we should
show viewDate as label there

We keep changeDayLabel in DateInput since
it needs to be in the input

Add label for Calendar in DateInput

Fix attribute name

Add accessible prop

Add a11y section into DateInput doc

Add calendar toggle on calendar icon click

Open calendar on Click on text

Don't show calendar icon if calendar show is off

Open calendar when user click on input

CalendarRange: Small fix for keyboard Enter/Space click behavior

Allow to control restoreFocus of FocusTrap

Move Calendar back inside FormField

Add note that calendar is now can be opened by click on input