Remove dispatching of enableKeyboardInput event
In CalendarDay on focus we used to dispatch this event
but it leads to an issue.
Whe user clicks on the day via mouse that closes the calendar
and moves focus back to previously focused element,
but since we dispatch that event we focused element
visual focus is visible but it should not be visible as we
used mouse.
In case of the keyboard it would make sence